+++ src/pynwb/ecephys.py
@@ -4,12 +4,22 @@ from collections.abc import Iterable
 
 from hdmf.common import DynamicTableRegion
 from hdmf.data_utils import assertEqualShape
-from hdmf.utils import docval, popargs, get_docval, popargs_to_dict, get_data_shape
+from hdmf.utils import docval, popargs, get_docval, popargs_to_dict, get_data_shape, AllowPositional
 
 from . import register_class, CORE_NAMESPACE
 from .base import TimeSeries
 from .core import NWBContainer, NWBDataInterface, MultiContainerInterface
 from .device import Device
+
+__all__ = [
+    'ElectrodeGroup',
+    'ElectricalSeries',
+    'SpikeEventSeries',
+    'EventDetection',
+    'LFP',
+    'FilteredEphys',
+    'FeatureExtraction'
+]
 
 
 @register_class('ElectrodeGroup', CORE_NAMESPACE)
@@ -29,7 +39,8 @@ class ElectrodeGroup(NWBContainer):
             {'name': 'position', 'type': 'array_data',
              'doc': 'Compound dataset with stereotaxic position of this electrode group (x, y, z). '
                     'The data array must have three elements or the dtype of the '
-                    'array must be ``(float, float, float)``', 'default': None})
+                    'array must be ``(float, float, float)``', 'default': None},
+            allow_positional=AllowPositional.WARNING,)
     def __init__(self, **kwargs):
         args_to_set = popargs_to_dict(('description', 'location', 'device', 'position'), kwargs)
         super().__init__(**kwargs)
@@ -91,7 +102,8 @@ class ElectricalSeries(TimeSeries):
              "filter is unknown, then this value could be 'Low-pass filter at 300 Hz'. If a non-standard filter "
              "type is used, provide as much detail about the filter properties as possible.", 'default': None},
             *get_docval(TimeSeries.__init__, 'resolution', 'conversion', 'timestamps', 'starting_time', 'rate',
-                        'comments', 'description', 'control', 'control_description', 'offset'))
+                        'comments', 'description', 'control', 'control_description', 'offset'),
+            allow_positional=AllowPositional.WARNING,)
     def __init__(self, **kwargs):
         args_to_set = popargs_to_dict(('electrodes', 'channel_conversion', 'filtering'), kwargs)
 
@@ -133,7 +145,8 @@ class SpikeEventSeries(ElectricalSeries):
              'doc': 'Timestamps for samples stored in data'},
             *get_docval(ElectricalSeries.__init__, 'electrodes'),  # required
             *get_docval(ElectricalSeries.__init__, 'resolution', 'conversion', 'comments', 'description', 'control',
-                        'control_description', 'offset'))
+                        'control_description', 'offset'),
+            allow_positional=AllowPositional.WARNING,)
     def __init__(self, **kwargs):
         data = kwargs['data']
         timestamps = kwargs['timestamps']
@@ -174,7 +187,8 @@ class EventDetection(NWBDataInterface):
                     '(e.g., .25msec before action potential peak, zero-crossing time, etc). '
                     'The index points to each event from the raw data'},
             {'name': 'times', 'type': ('array_data', 'data'), 'doc': 'Timestamps of events, in Seconds'},
-            {'name': 'name', 'type': str, 'doc': 'the name of this container', 'default': 'EventDetection'})
+            {'name': 'name', 'type': str, 'doc': 'the name of this container', 'default': 'EventDetection'},
+            allow_positional=AllowPositional.WARNING,)
     def __init__(self, **kwargs):
         args_to_set = popargs_to_dict(('detection_method', 'source_electricalseries', 'source_idx', 'times'), kwargs)
         super().__init__(**kwargs)
@@ -233,7 +247,9 @@ class Clustering(NWBDataInterface):
              'shape': (None,)},
             {'name': 'name', 'type': str, 'doc': 'the name of this container', 'default': 'Clustering'})
     def __init__(self, **kwargs):
-        warnings.warn("use pynwb.misc.Units or NWBFile.units instead", DeprecationWarning)
+        self._error_on_new_pass_on_construct(
+            error_msg='The Clustering neurodata type is deprecated. Use pynwb.misc.Units or NWBFile.units instead'
+        )
         args_to_set = popargs_to_dict(('description', 'num', 'peak_over_rms', 'times'), kwargs)
         super().__init__(**kwargs)
         args_to_set['peak_over_rms'] = list(args_to_set['peak_over_rms'])
@@ -267,7 +283,9 @@ class ClusterWaveforms(NWBDataInterface):
              'doc': 'the standard deviations of waveforms for each cluster'},
             {'name': 'name', 'type': str, 'doc': 'the name of this container', 'default': 'ClusterWaveforms'})
     def __init__(self, **kwargs):
-        warnings.warn("use pynwb.misc.Units or NWBFile.units instead", DeprecationWarning)
+        self._error_on_new_pass_on_construct(
+            error_msg='The ClusterWaveforms neurodata type is deprecated. Use pynwb.misc.Units or NWBFile.units instead'
+        )
         args_to_set = popargs_to_dict(('clustering_interface', 'waveform_filtering',
                                        'waveform_mean', 'waveform_sd'), kwargs)
         super().__init__(**kwargs)
@@ -333,7 +351,8 @@ class FeatureExtraction(NWBDataInterface):
              'doc': 'The times of events that features correspond to'},
             {'name': 'features', 'type': ('array_data', 'data'), 'shape': (None, None, None),
              'doc': 'Features for each channel'},
-            {'name': 'name', 'type': str, 'doc': 'the name of this container', 'default': 'FeatureExtraction'})
+            {'name': 'name', 'type': str, 'doc': 'the name of this container', 'default': 'FeatureExtraction'},
+            allow_positional=AllowPositional.WARNING,)
     def __init__(self, **kwargs):
         # get the inputs
         electrodes, description, times, features = popargs(